This update mostly improves synchronization (but not only)
- Fixes the calculation of speeds below 1.0, which could be falsified by
rounding off some decimal.
- Adds a per world "sync" option that avoids tick offsets. When
activated, it will keep the tick constantly synchronized with the
server. It prevent any "/tm set time" command and replace it by a "/tm
set start", sleep until the next dawn will become impossible, warning
messages are displayed and/or related options are automatically changed.
- Lowers the minimum refresh rate to 2 ticks, it seems to work properly.
- Adds a public debug mode, which can be activated through the
'debugMode' option in the config file or the related command. It will
display colored messages in the console, allowing for more accurate
reporting.
- Server version detection is upgraded too, so it should close issue 
https://github.com/ArVdC/TimeManager/issues/1
- Renames and reworks some sub commands and related classes and options,
for a more convenient usage, especially when using the tab completer.
- Now if a player omits to specify the 'world' argument in a command, it
will be replaced by the world in which it stands.
- Tab completer has been upgraded, but still contains a hack for
managing spaces in world names.