This is the 1.1.1 release, ready for the mc 1.13 update :)
* Lib change : Now using the mc 1.13 library (the 'setGameRuleValue' was
actualized)
* Cmd change : '/tm checktimers' became '/tm checktime
[world|server|all]' and allows admins to target the parameters they want
to check.
* New cmd : '/tm checkconfig' displays a summary of the configuration
file.
* New cmd : '/tm set initialtick' changes the initialTickNb value in the
configuration file and the MySQL database if it is active.
* OnReload changes :
	* Now, manually change the initialTickNb in the yaml file could
updates the value in the MySQL database and forces a resynchronization
of all the timers.
	* If someone on your network has changed the initialTickNb in the MySQL
database, the change will immediately be reflected on your server when
the configuration is reloaded.
* New scheduleSyncDelayedTask : If someone on your network has changed
the initialTickNb in the MySQL database, the change will be reflected on
every other servers within the next 2 minutes.